+++ jog_arm/CMakeLists.txt
@@ -31,7 +31,12 @@ include_directories(
   ${Eigen_INCLUDE_DIRS}
 )
 
-add_executable(jog_arm_server src/jog_arm/jog_arm_server.cpp)
+add_executable(jog_arm_server
+        src/jog_arm/jog_arm_server.cpp
+        src/jog_arm/jog_ros_interface.cpp
+        src/jog_arm/low_pass_filter.cpp
+        src/jog_arm/jog_calcs.cpp
+        src/jog_arm/collision_check_thread.cpp)
 add_dependencies(jog_arm_server ${catkin_EXPORTED_TARGETS})
 target_link_libraries(jog_arm_server ${catkin_LIBRARIES} ${Eigen_LIBRARIES})
